+++ OpenDreamRuntime/DreamManager.cs
@@ -213,7 +213,7 @@ public sealed partial class DreamManager {
                 _atomManager.TryGetMovableFromEntity(reference.Entity, out var atom);
                 return atom;
             case ClientObjectReference.RefType.Turf:
-                _dreamMapManager.TryGetTurfAt(new((uint)reference.TurfX, (uint)reference.TurfY), (uint)reference.TurfZ, out var turf);
+                _dreamMapManager.TryGetTurfAt(new(reference.TurfX, reference.TurfY), reference.TurfZ, out var turf);
                 return turf;
         }
